src/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformers
timm>=0.9.0
transformers>=4.30.0

# Explainability
pytorch-grad-cam>=1.4.0
scipy>=1.10.0

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Federated Learning & Blockchain
web3>=6.0.0
eth-account>=0.8.0
eth-tester>=0.9.0
py-solc-x>=1.1.0
py-evm>=0.7.0

# Web UI (Optional)
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# YOLO Integration (for object detection-based X-ray analysis)
ultralytics>=8.0.0
opencv-python>=4.8.0
